# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wolt API",
        "description": "API documentation for the Wolt-like application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Create an order for a restaurant",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "example": "user-session-12345"
                                    },
                                    "restaurant_id": {
                                        "type": "string",
                                        "example": "677993288fbedfb0b45059fd"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_item_id": {
                                                    "type": "string",
                                                    "example": "Kanapka Norweska"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurant not found"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "description": "Fetch details of an existing order by ID",
                "operationId": "getOrderDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "description": "Update the status of an existing order",
                "operationId": "updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "description": "Cancel an existing order by ID",
                "operationId": "cancelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "string"
                    },
                    "session_id": {
                        "description": "Session ID",
                        "type": "string"
                    },
                    "restaurant_id": {
                        "description": "Restaurant ID",
                        "type": "string"
                    },
                    "items": {
                        "description": "Ordered items",
                        "type": "array",
                        "items": {
                            "properties": {
                                "menu_item_id": {
                                    "description": "Menu item ID",
                                    "type": "string"
                                },
                                "quantity": {
                                    "description": "Quantity ordered",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "total_price": {
                        "description": "Total price of the order",
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "description": "Order status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "created_at": {
                        "description": "Order creation time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Order last update time",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}